ログアウト時に＋を非表示
